Scale up the forest (8 trees, depth 14 = 131,064 segments)
Bigger workload so the GPU's parallelism clearly dominates and the device->host
copy stops being a meaningful share of the frame:

- forest_schema.py: TREES 3->8, TREE_DEPTH 12->14, SEG_CAP 12288->131072,
  WORLD 1280x800 -> 1920x1000. Region grows 627KB -> 6.7MB (a fixed-layout size
  change — regenerated with --allow-breaking; single-version demo).
- producer.mojo: array-length setters are now u32 (cap > 65535).
- renderer.odin: UNCHANGED — it reads WORLD_W/H and TREE_DEPTH from the overlay's
  schema constants, so the bump propagated automatically.
- README + GIF refreshed for the larger, denser forest.

Measured on M4 Max (per-frame compute, sleep removed): CPU ~8.6 ms (116 fps) vs GPU
~0.44 ms (2280 fps) = ~19x (was ~1.9x at 12k). Renderer draws 131k DrawLineEx at
~84 fps. Everything still clears 60 fps.

+++ producer.mojo
@@ -102,11 +102,11 @@ def write_branch(f: Forest, g: Int, time: Float32):
 def set_headers(f: Forest) raises:
     # count + array lengths — constant every frame; the geometry is what changes.
     f.set_count(UInt32(TOTAL))
-    f.set_x0_len(UInt16(TOTAL))
-    f.set_y0_len(UInt16(TOTAL))
-    f.set_x1_len(UInt16(TOTAL))
-    f.set_y1_len(UInt16(TOTAL))
-    f.set_depth_len(UInt16(TOTAL))
+    f.set_x0_len(UInt32(TOTAL))
+    f.set_y0_len(UInt32(TOTAL))
+    f.set_x1_len(UInt32(TOTAL))
+    f.set_y1_len(UInt32(TOTAL))
+    f.set_depth_len(UInt32(TOTAL))
 
 
 # ── GPU kernel: one thread per branch ─────────────────────────────────────────
